Ideal Index.mjs Layout
The function for the main menu has been seperate from the other sub-menus. These allow recursion (one can re-enter the same sub-menu from that sub-menu) and displaying a table does not get in the way thanks to an async/await. Each sub-menu can display all their data, now to fill in the rest of CRUD